rename golfGame->golf_game
